Strengthen tests for the monitoring.max-concurrency limiter
Move the concurrency test into monitoring_collector_test.go (the
corresponding _test.go file for the changed code), extract semaphore
construction into a small newRequestLimiter helper with its own
table-driven test, and add direct unit tests for acquire/releaseRequestLimiter
covering the nil (unbounded) and blocking-when-full cases. Also assert
Config.MaxConcurrentRequests defaults correctly in TestNewConfigWithDefaults.
